+++ src/Data/NBB.Data.Abstractions/IUowRepository.cs
@@ -1,4 +1,6 @@
-﻿using NBB.Core.Abstractions;
+﻿using System.Threading;
+using System.Threading.Tasks;
+using NBB.Core.Abstractions;
 
 namespace NBB.Data.Abstractions
 {
@@ -6,5 +8,14 @@ namespace NBB.Data.Abstractions
         where TEntity : class
     {
         IUow<TEntity> Uow { get; }
+    }
+    
+    public static class UowRepositoryExtensions
+    {
+        public static Task SaveChangesAsync<TEntity>(this IUowRepository<TEntity> repository, CancellationToken cancellationToken = default)
+            where TEntity : class
+        {
+            return repository.Uow.SaveChangesAsync(cancellationToken);
+        }
     }
 }
